+++ src/syntax.js
@@ -168,6 +168,8 @@
     
                     copyButton.onclick = function() {
                         _parameter_Navigator.clipboard.writeText( innerHTMLCopy );
+
+                        fireCustomTrigger( "onCopy", innerHTMLCopy );
                     };
                 }