+++ docs/CHANGE_LOG.md
@@ -3,6 +3,7 @@
 ## Version 0.3.0:
 - Added Code/Pre element rendering support.
 - Added options attribute "data-syntax-options" support, which will override the default options per code element rendered.
+- Fixed a class name setting issue that caused a random space to appear.
 - 
 
 <br>